// apps/web/src/api/models/getOrderResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Autodor.OpenApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { OrderItemResponse } from './orderItemResponse';

export interface GetOrderResponse {
  id: string;
  /** @nullable */
  number: string | null;
  date: string;
  /** @nullable */
  person: string | null;
  /** @nullable */
  customerNumber: string | null;
  items: OrderItemResponse[];
  isExcluded: boolean;
}